Stop Generation & Aggressive reload of model (#16)
* Stop generation of a model Solves a problem where if an app goes into
the background, you have no way to stop generation and this can lead to
a crash of llama.cpp
* Check model status before starting generation Added the ability to
check the model loaded status directly from the llama.cpp lib rather
than store it in a variable. iOS has a nasty habit of unloading things
from memory when the app goes into the background, so this is a more
reliable way to check if the model is loaded. Additionally, the client
will now agressively try to reload the model if it is not loaded when
starting generation.