Filter out header only assistant message (#34)
This pull request introduces a small but important change to the
`generateCloudChatCompletion` method in
`Sources/FreeToken/FreeToken.swift`. The update ensures that empty
assistant messages are filtered out before processing, which is not
required for cloud completions.

*
[`Sources/FreeToken/FreeToken.swift`](diffhunk://#diff-6d471ed24ea5a6c41c69766580f15aa6739a16a1be7369ff8892db13f9187d29R786-R789):
Added a filter to remove empty assistant messages from the `messages`
array in the `generateCloudChatCompletion` method.