Configuration of batch size & stop token changes (#41)
This pull request introduces changes to replace the `stopTokens`
property with a new `batchSize` property across multiple files in the
`FreeToken` module. Additionally, platform-specific client configuration
logic has been added for macOS and iOS. These changes enhance the
flexibility of the AI model configuration and improve platform-specific
support.

### Replacement of `stopTokens` with `batchSize`:

*
[`Sources/FreeToken/AIModelConfiguration.swift`](diffhunk://#diff-37a373e495d5720f75b01a75a99298991d5f1a07e69ab38acb7667f6909e2201L14-R14):
Replaced the `stopTokens` property with `batchSize` in the `FreeToken`
model configuration and updated the initializer to handle this change.
[[1]](diffhunk://#diff-37a373e495d5720f75b01a75a99298991d5f1a07e69ab38acb7667f6909e2201L14-R14)
[[2]](diffhunk://#diff-37a373e495d5720f75b01a75a99298991d5f1a07e69ab38acb7667f6909e2201L27-R27)
*
[`Sources/FreeToken/AIModelManager.swift`](diffhunk://#diff-67836cde4e812c9f6b06fc38c440e02bc1e7a09ce157d35c6f39856d898f3855R173):
Updated the AI model manager to use `batchSize` instead of `stopTokens`
when initializing model parameters and configurations.
[[1]](diffhunk://#diff-67836cde4e812c9f6b06fc38c440e02bc1e7a09ce157d35c6f39856d898f3855R173)
[[2]](diffhunk://#diff-67836cde4e812c9f6b06fc38c440e02bc1e7a09ce157d35c6f39856d898f3855L251-R252)
*
[`Sources/FreeToken/Resources.swift`](diffhunk://#diff-f23cca06ce81e4c66d12ff385b29ce69764c37b39dbfcc3d0b056ff4fb7f854eL128-R128):
Updated the coding keys to map `batchSize` to the corresponding JSON
field and removed references to `stopTokens`.
[[1]](diffhunk://#diff-f23cca06ce81e4c66d12ff385b29ce69764c37b39dbfcc3d0b056ff4fb7f854eL128-R128)
[[2]](diffhunk://#diff-f23cca06ce81e4c66d12ff385b29ce69764c37b39dbfcc3d0b056ff4fb7f854eL140-R140)

### Platform-specific client configuration:

*
[`Sources/FreeToken/AIModelManager.swift`](diffhunk://#diff-67836cde4e812c9f6b06fc38c440e02bc1e7a09ce157d35c6f39856d898f3855R342-R346):
Added conditional logic to set the `clientConfig` based on the operating
system (macOS or iOS). This ensures proper client configuration for each
platform.