Highlander mode (#77)
Highlander mode is on iOS there can "only be one ⚡️🗡️" instance of an AI
model running on the client. MacOS does not run highlander mode and will
allow multiple instances to load onto the GPU as memory allows (based on
the model).

This has brought a lot of stability but trades off loading instances
into KV cache on the GPU.